Add files via upload
This is the PDF version of the actual Cyclistic Dashboard. As the size of the actual tableau workbook file is more than 25MB, it was not allowed to upload to Github. I have included a link within the PDF file, directing to the Tableau Public website, where you can find this Dashboard in Tableau Workbook format.